requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
backoff==2.2.1
billiard==4.2.0
blurhash-python==1.2.2
botocore==1.34.58
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-graphql-jwt==0.4.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
dnspython==2.6.1
email_validator==2.1.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.3.1
gql==3.5.0
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
kombu==5.3.5
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
mypy==1.7.1
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
pillow==10.2.0
promise==2.3
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
strawberry-django-auth==0.376.5
strawberry-graphql==0.219.2
strawberry-graphql-django==0.32.2
sympy==1.12
text-unidecode==1.3
torch==2.2.1
torchvision==0.17.1
triton==2.2.0
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
websockets==11.0.3
yarl==1.9.4
